+++ src/flows/CostCalculator/index.ts
@@ -1,3 +1,2 @@
 export { CostCalculator } from './CostCalculator';
-export { useCostCalculator } from './hooks';
-export { useExportAsPdf } from './exportAsPdf';
+export { useCostCalculator, useExportAsPdf } from './hooks';
