+++ src/index.tsx
@@ -2,7 +2,7 @@ import './styles/global.css';
 
 export { CostCalculator } from '@/src/flows/CostCalculator';
 // eslint-disable-next-line react-refresh/only-export-components
-export { useExportAsPdf } from '@/src/flows/CostCalculator/exportAsPdf';
+export { useExportAsPdf } from '@/src/flows/CostCalculator/hooks';
 export { RemoteFlows } from '@/src/RemoteFlowsProvider';
 export { ThemeProvider } from '@/src/theme';
 // eslint-disable-next-line react-refresh/only-export-components
